# swagger.json
{"swagger":"2.0","info":{"description":"","version":"1.0.0","title":"Swagger 接口文档"},"host":"localhost:3000","basePath":"","tags":[{"name":"get","description":"GET 请求测试","externalDocs":{"description":"","url":""}},{"name":"jwt","description":"jwt 验证","externalDocs":{"description":"","url":""}},{"name":"oauth","description":"oauth 验证","externalDocs":{"description":"","url":""}},{"name":"login","description":"POST 请求测试","externalDocs":{"description":"","url":""}}],"schemes":["http"],"paths":{"/get/test":{"get":{"tags":["get"],"summary":"接口标题","description":"接口描述","operationId":"0.20061696202042278","consumes":["application/json"],"produces":["application/json"],"responses":{},"security":[{"petstore_auth":["write:pets","read:pets"]},{"api_key":[]}],"deprecated":false}},"/get/test1":{"get":{"tags":["get"],"summary":"接口标题","description":"接口描述","operationId":"0.24681384033458365","consumes":["application/json"],"produces":["application/json"],"responses":{},"security":[{"petstore_auth":["write:pets","read:pets"]},{"api_key":[]}],"deprecated":false},"post":{"tags":["get"],"summary":"接口标题","description":"接口描述","operationId":"0.0007369547859190639","consumes":["application/json"],"produces":["application/json"],"responses":{},"security":[{"petstore_auth":["write:pets","read:pets"]},{"api_key":[]}],"deprecated":false}},"/jwt/login-token":{"post":{"tags":["jwt"],"summary":"接口标题","description":"接口描述","operationId":"0.6464588484088034","consumes":["application/json"],"produces":["application/json"],"responses":{},"security":[{"petstore_auth":["write:pets","read:pets"]},{"api_key":[]}],"deprecated":false}},"/jwt/getUser-token":{"get":{"tags":["jwt"],"summary":"接口标题","description":"接口描述","operationId":"0.05585377434254313","consumes":["application/json"],"produces":["application/json"],"responses":{},"security":[{"petstore_auth":["write:pets","read:pets"]},{"api_key":[]}],"deprecated":false}},"/oauth/github/login":{"get":{"tags":["oauth"],"summary":"接口标题","description":"接口描述","operationId":"0.08123405976642006","consumes":["application/json"],"produces":["application/json"],"responses":{},"security":[{"petstore_auth":["write:pets","read:pets"]},{"api_key":[]}],"deprecated":false}},"/oauth/github/callback":{"get":{"tags":["oauth"],"summary":"接口标题","description":"接口描述","operationId":"0.3364128272318547","consumes":["application/json"],"produces":["application/json"],"responses":{},"security":[{"petstore_auth":["write:pets","read:pets"]},{"api_key":[]}],"deprecated":false}},"/login/":{"post":{"tags":["login"],"summary":"接口标题","description":"接口描述","operationId":"0.39830621044195347","consumes":["application/json"],"produces":["application/json"],"responses":{},"security":[{"petstore_auth":["write:pets","read:pets"]},{"api_key":[]}],"deprecated":false}}},"securityDefinitions":{"petstore_auth":{"type":"oauth2","authorizationUrl":"http://petstore.swagger.io/oauth/dialog","flow":"implicit","scopes":{"write:pets":"modify pets in your account","read:pets":"read your pets"}},"api_key":{"type":"apiKey","name":"api_key","in":"header"}},"definitions":{"Order":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"petId":{"type":"integer","format":"int64"},"quantity":{"type":"integer","format":"int32"},"shipDate":{"type":"string","format":"date-time"},"status":{"type":"string","description":"Order Status","enum":["placed","approved","delivered"]},"complete":{"type":"boolean","default":false}},"xml":{"name":"Order"}},"Category":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}},"xml":{"name":"Category"}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"phone":{"type":"string"},"userStatus":{"type":"integer","format":"int32","description":"User Status"}},"xml":{"name":"User"}},"Tag":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}},"xml":{"name":"Tag"}},"Pet":{"type":"object","required":["name","photoUrls"],"properties":{"id":{"type":"integer","format":"int64"},"category":{"$ref":"#/definitions/Category"},"name":{"type":"string","example":"doggie"},"photoUrls":{"type":"array","xml":{"name":"photoUrl","wrapped":true},"items":{"type":"string"}},"tags":{"type":"array","xml":{"name":"tag","wrapped":true},"items":{"$ref":"#/definitions/Tag"}},"status":{"type":"string","description":"pet status in the store","enum":["available","pending","sold"]}},"xml":{"name":"Pet"}},"ApiResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"type":{"type":"string"},"message":{"type":"string"}}}}}